requirements.txt
# 核心依赖
python-docx>=1.1.0
pyyaml>=6.0
httpx>=0.27.0
click>=8.1.0
pydantic>=2.0.0

# Web图形界面依赖
Flask>=3.0.0
PyQt6>=6.6.0
PyQt6-WebEngine>=6.6.0
